+++ src/routes/privacyPolicy.ts
@@ -1,0 +1,5 @@
+module.exports = (app: any) => {
+    app.get('/privacy-policy', (req: any, res:any) => {
+        res.status(301).redirect('https://drive.google.com/a/imsa.edu/file/d/11_cAuaerCrQ3BBXNx_G_zw1ZyGaTWx0z/view?usp=sharing');
+    });
+}
